refactor: move plugin to core2
